+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/resource/kubernetes/PodOperator.java
@@ -4,6 +4,7 @@
  */
 package io.strimzi.operator.cluster.operator.resource.kubernetes;
 
+import io.fabric8.kubernetes.api.model.DeletionPropagation;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodList;
@@ -24,6 +25,8 @@ public class PodOperator extends AbstractReadyNamespacedResourceOperator<Kuberne
     private static final ReconciliationLogger LOGGER = ReconciliationLogger.create(PodOperator.class);
     private static final String NO_UID = "NULL";
 
+    private boolean useBackgroundDeletion = false;
+
     /**
      * Constructor
      * @param vertx The Vertx instance
@@ -31,6 +34,22 @@ public class PodOperator extends AbstractReadyNamespacedResourceOperator<Kuberne
      */
     public PodOperator(Vertx vertx, KubernetesClient client) {
         super(vertx, client, "Pods");
+    }
+
+    /**
+     * Constructor
+     * @param vertx                 The Vertx instance
+     * @param client                The Kubernetes client
+     * @param useBackgroundDeletion Whether to use background deletion propagation when deleting pods
+     */
+    public PodOperator(Vertx vertx, KubernetesClient client, boolean useBackgroundDeletion) {
+        super(vertx, client, "Pods");
+        this.useBackgroundDeletion = useBackgroundDeletion;
+    }
+
+    @Override
+    protected DeletionPropagation getDeletionPropagation(boolean cascading) {
+        return useBackgroundDeletion ? DeletionPropagation.BACKGROUND : super.getDeletionPropagation(cascading);
     }
 
     @Override