+++ operator-common/src/main/java/io/strimzi/operator/common/featuregates/FeatureGates.java
@@ -26,11 +26,14 @@ public class FeatureGates {
     private static final String SERVER_SIDE_APPLY_PHASE_1 = "ServerSideApplyPhase1";
     // Enables the usage of Buildah in Connect Build
     private static final String USE_CONNECT_BUILD_WITH_BUILDAH = "UseConnectBuildWithBuildah";
+    // Enables background deletion propagation when rolling pods
+    private static final String USE_BACKGROUND_POD_DELETION = "UseBackgroundPodDeletion";
 
     // When adding new feature gates, do not forget to add them to allFeatureGates(), toString(), equals(), and `hashCode() methods
     private final FeatureGate serverSideApplyPhase1 = new FeatureGate(SERVER_SIDE_APPLY_PHASE_1, true);
     // private final FeatureGate dummyFeatureGate = new FeatureGate(DUMMY_FEATURE_GATE, false);
     private final FeatureGate useConnectBuildWithBuildah = new FeatureGate(USE_CONNECT_BUILD_WITH_BUILDAH, true);
+    private final FeatureGate useBackgroundPodDeletion = new FeatureGate(USE_BACKGROUND_POD_DELETION, false);
 
     /**
      * Constructs the feature gates configuration.
@@ -60,6 +63,9 @@ public class FeatureGates {
                         break;
                     case USE_CONNECT_BUILD_WITH_BUILDAH:
                         setValueOnlyOnce(useConnectBuildWithBuildah, value);
+                        break;
+                    case USE_BACKGROUND_POD_DELETION:
+                        setValueOnlyOnce(useBackgroundPodDeletion, value);
                         break;
                     default:
                         throw new InvalidConfigurationException("Unknown feature gate " + featureGate + " found in the configuration");
@@ -117,6 +123,15 @@ public class FeatureGates {
     }
 
     /**
+     * Checks if the UseBackgroundPodDeletion feature gate is enabled.
+     *
+     * @return  Returns true when the UseBackgroundPodDeletion feature gate is enabled
+     */
+    public boolean useBackgroundPodDeletionEnabled() {
+        return useBackgroundPodDeletion.isEnabled();
+    }
+
+    /**
      * Returns a list of all Feature gates. Used for testing.
      *
      * @return  List of all Feature Gates
@@ -125,7 +140,8 @@ public class FeatureGates {
         return List.of(
         //  dummyFeatureGate
             serverSideApplyPhase1,
-            useConnectBuildWithBuildah
+            useConnectBuildWithBuildah,
+            useBackgroundPodDeletion
         );
     }
 
@@ -134,7 +150,8 @@ public class FeatureGates {
         return "FeatureGates(" +
     //      "DummyFeatureGate=" + dummyFeatureGate.isEnabled() +
             "ServerSideApplyPhase1=" + serverSideApplyPhase1.isEnabled() + ", " +
-            "UseConnectBuildWithBuildah=" + useConnectBuildWithBuildah.isEnabled() +
+            "UseConnectBuildWithBuildah=" + useConnectBuildWithBuildah.isEnabled() + ", " +
+            "UseBackgroundPodDeletion=" + useBackgroundPodDeletion.isEnabled() +
             ")";
     }
 
@@ -170,13 +187,14 @@ public class FeatureGates {
             FeatureGates other = (FeatureGates) o;
             // return Objects.equals(dummyFeatureGate, other.dummyFeatureGate)
             return Objects.equals(serverSideApplyPhase1, other.serverSideApplyPhase1)
-                && Objects.equals(useConnectBuildWithBuildah, other.useConnectBuildWithBuildah);
+                && Objects.equals(useConnectBuildWithBuildah, other.useConnectBuildWithBuildah)
+                && Objects.equals(useBackgroundPodDeletion, other.useBackgroundPodDeletion);
         }
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(serverSideApplyPhase1, useConnectBuildWithBuildah);
+        return Objects.hash(serverSideApplyPhase1, useConnectBuildWithBuildah, useBackgroundPodDeletion);
     }
 
     /**